refactor(mall-distribution-aplication):
修改了违反结构单一性原则的一条代码
