+++ mall-distribution-application/src/main/java/com/mall/distribution/application/port/CommissionCashCreateUseCase.java
@@ -1,0 +1,10 @@
+package com.mall.distribution.application.port;
+
+import com.mall.distribution.application.command.CreateCommissionCashCommand;
+
+public interface CommissionCashCreateUseCase {
+    /**
+     * 创建佣金提现申请
+     */
+    Long createCommissionCash(CreateCommissionCashCommand command);
+}
