+++ mall-distribution-application/src/main/java/com/mall/distribution/application/service/CommissionCashCreateUseCaseService.java
@@ -1,0 +1,11 @@
+package com.mall.distribution.application.service;
+
+import com.mall.distribution.application.command.CreateCommissionCashCommand;
+import com.mall.distribution.application.port.CommissionCashCreateUseCase;
+
+public class CommissionCashCreateUseCaseService implements CommissionCashCreateUseCase {
+    @Override
+    public Long createCommissionCash(CreateCommissionCashCommand command) {
+        return null;
+    }
+}
