+++ mall-distribution-application/src/main/java/com/mall/distribution/application/service/CommissionCashService.java
@@ -4,14 +4,15 @@ import com.mall.distribution.application.command.CreateCommissionCashCommand;
 import com.mall.distribution.application.port.CommissionCashUseCase;
 import com.mall.distribution.application.port.DistributorUseCase;
 import com.mall.distribution.application.query.CommissionCashQuery;
+import com.mall.distribution.common.enums.Resp;
 import com.mall.distribution.common.exception.BusinessException;
-import com.mall.distribution.common.exception.Resp;
-import com.mall.distribution.domain.commission.CommissionCash;
-import com.mall.distribution.domain.commission.CommissionCashRepository;
-import com.mall.distribution.domain.distributor.Distributor;
+
+import com.mall.distribution.domain.model.cash.CommissionCash;
+import com.mall.distribution.domain.model.distributor.Distributor;
+import com.mall.distribution.domain.repository.CommissionCashRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.List;
 
@@ -21,21 +22,21 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class CommissionCashService implements CommissionCashUseCase {
-    
+
     private final CommissionCashRepository commissionCashRepository;
     private final DistributorUseCase distributorUseCase;
-    
+
     @Override
     @Transactional
     public Long createCommissionCash(CreateCommissionCashCommand command) {
         // 查找分销商
         Distributor distributor = distributorUseCase.findById(command.getDistributorId());
-        
+
         // 检查佣金余额是否足够
         if (distributor.getCommissionBalance().compareTo(command.getCashAmount()) < 0) {
             throw new BusinessException(Resp.COMMISSION_BALANCE_INSUFFICIENT);
         }
-        
+
         // 创建提现申请
         CommissionCash commissionCash = CommissionCash.create(
             command.getDistributorId(),
@@ -45,24 +46,24 @@ public class CommissionCashService implements CommissionCashUseCase {
             command.getBankBranch(),
             command.getCreatedBy()
         );
-        
+
         // 保存提现申请
         commissionCashRepository.save(commissionCash);
-        
+
         return commissionCash.getCashId();
     }
-    
+
     @Override
     public CommissionCash findById(Long cashId) {
         return commissionCashRepository.findById(cashId)
             .orElseThrow(() -> new BusinessException(Resp.COMMISSION_CASH_NOT_FOUND));
     }
-    
+
     @Override
     public List<CommissionCash> findCommissionCashes(CommissionCashQuery query) {
         return commissionCashRepository.findByQuery(query);
     }
-    
+
     @Override
     @Transactional
     public void approveCash(Long cashId, String auditBy, String auditRemark) {
@@ -70,7 +71,7 @@ public class CommissionCashService implements CommissionCashUseCase {
         commissionCash.approve(auditBy, auditRemark);
         commissionCashRepository.save(commissionCash);
     }
-    
+
     @Override
     @Transactional
     public void rejectCash(Long cashId, String auditBy, String auditRemark) {
@@ -78,18 +79,18 @@ public class CommissionCashService implements CommissionCashUseCase {
         commissionCash.reject(auditBy, auditRemark);
         commissionCashRepository.save(commissionCash);
     }
-    
+
     @Override
     @Transactional
     public void payCash(Long cashId, String paymentNo, String paidBy) {
         CommissionCash commissionCash = findById(cashId);
         commissionCash.pay(paymentNo, paidBy);
         commissionCashRepository.save(commissionCash);
-        
+
         // 扣减分销商佣金余额
         distributorUseCase.updateCommissionBalance(
-            commissionCash.getDistributorId(), 
-            commissionCash.getCashAmount().negate(), 
+            commissionCash.getDistributorId(),
+            commissionCash.getCashAmount().negate(),
             paidBy
         );
     }
